refactor(schlib): split primitives.rs into a primitives/ module directory
Pure code motion (U1 of the SchLib/PcbLib structural unification): the
monolithic primitives.rs becomes primitives/{mod,pins,shapes,text,models}.rs,
mirroring pcblib/primitives/. All public paths are preserved via re-exports
in primitives/mod.rs — no file outside the module changed. No behaviour
change; gates green incl. the readability oracle (0 regressions).
